(#89) ♻️: GenerateMyIdUseCase를 RegisterUserUseCase로 리팩터링
